requirements_local.txt
# local environment requirements

#includes the base.txt requirements needed in all environments
-r requirements.txt

# below are requirements specific to the local environment
django-sslserver
django-debug-toolbar